WS1: extract ClusteredMap renderer into clustered-map.tsx
Fifth slice of the modelmap.tsx split (913 -> 451). Move the ~440-line
ClusteredMap component (MapLibre sources/layers, cluster + people/cases dot
rendering, popups, marker click handling, fit-bounds, EmojiOverlay mount) into
src/features/model-map/clustered-map.tsx as a default-exported client component.
modelmap.tsx now holds only the ModelMap orchestrator (data loading, playback,
mode/timestep state) and renders <ClusteredMap />.

Verified: tsc clean, lint clean, pnpm test 23/23, and the run page renders the
map with zero console errors (browser smoke test on run 30).
